+++ lib/core/facets/kernel/callstack.rb
@@ -3,12 +3,7 @@ module Kernel
   alias_method :pp_callstack, :caller
   alias_method :pp_call_stack, :caller
 
-  # Parse a caller string and break it into its components,
-  # returning an array composed of:
-  #
-  # * file (String)
-  # * lineno (Integer)
-  # * method (Symbol)
+  # Returns the call stack as an array of [file, lineno, method] entries.
   #
   # For example, from irb
   #
@@ -16,27 +11,14 @@ module Kernel
   #
   # _produces_ ...
   #
-  #     [["(irb)", 2, :irb_binding],
-  #       ["/usr/lib/ruby/1.8/irb/workspace.rb", 52, :irb_binding],
-  #       ["/usr/lib/ruby/1.8/irb/workspace.rb", 52, nil]]
-  #
-  # Note: If the user decides to redefine caller() to output data
-  # in a different format, _prior_ to requiring this, then the
-  # results will be indeterminate.
+  #     [["(irb)", 2, :irb_binding], ...]
   #
   # CREDIT: Trans
 
   def callstack(level = 1)
-    call_str_array = pp_callstack(level)
-    stack = []
-    call_str_array.each{ |call_str|
-      file, lineno, method = call_str.split(':')
-      if method =~ /in [`'](.*)'/ then
-        method = $1.intern()
-      end
-      stack << [file, lineno.to_i, method]
-    }
-    stack
+    caller_locations(level).map do |loc|
+      [loc.path, loc.lineno, loc.label&.to_sym]
+    end
   end
 
   alias_method :call_stack, :callstack
@@ -53,4 +35,3 @@ class Binding
 
   alias_method :call_stack, :callstack
 end
-