+++ lib/core/facets/kernel/require_all.rb
@@ -1,23 +1,15 @@
 module Kernel
 
-  # Require a pattern of files relatvie to the current file.
-  # This makes is easy to require an entire directory, for instance:
+  # Require a pattern of files relative to the current file.
+  # This makes it easy to require an entire directory, for instance:
   #
   #   require_all 'core_ext/*'
   #
-  # NOTE: This method used to allow glob-based requires from the $LOAD_PATH,
-  # but this was deprecated in favor of relative requiring only, as it is
-  # consider the typical usecase, and globbing from the $LOAD_PATH is a
-  # bit dangerous. Better options exist for globbing the $LOAD_PATH such as
-  # the +plugins+ gem.
 
   def require_all(pattern)
-    c = caller.first
-    fail "Can't parse #{c}" unless c.rindex(/:\d+(:in [`'].*')?$/)
-    file = $` # File.dirname(c)
-    if /\A\((.*)\)/ =~ file # eval, etc.
-      raise LoadError, "require_relative is called in #{$1}"
-    end
+    loc = caller_locations(1, 1).first
+    file = loc.absolute_path || loc.path
+    raise LoadError, "require_all is called in #{loc.label}" unless file
     glob = File.expand_path(pattern, File.dirname(file))
     Dir.glob(glob).each do |absolute|
       require absolute
@@ -26,12 +18,9 @@ module Kernel
 
   # Same as #require_all, but for #load.
   def load_all(pattern, safe=nil)
-    c = caller.first
-    fail "Can't parse #{c}" unless c.rindex(/:\d+(:in [`'].*')?$/)
-    file = $` # File.dirname(c)
-    if /\A\((.*)\)/ =~ file # eval, etc.
-      raise LoadError, "require_relative is called in #{$1}"
-    end
+    loc = caller_locations(1, 1).first
+    file = loc.absolute_path || loc.path
+    raise LoadError, "load_all is called in #{loc.label}" unless file
     glob = File.expand_path(pattern, File.dirname(file))
     Dir.glob(glob).each do |absolute|
       load absolute, safe
@@ -39,4 +28,3 @@ module Kernel
   end
 
 end
-